+++ plugins/jaeger/src/components/page/TracesToolbarOperations.tsx
@@ -29,7 +29,7 @@ const TracesToolbarOperations: React.FunctionComponent<ITracesToolbarOperationsP
         const json = await response.json();
 
         if (response.status >= 200 && response.status < 300) {
-          return ['All Operations', ...json.data.map((operation: IOperation) => operation.name)];
+          return ['All Operations', ...json.data.map((operation: IOperation) => operation.name).sort()];
         } else {
           if (json.error) {
             throw new Error(json.error);
@@ -44,12 +44,13 @@ const TracesToolbarOperations: React.FunctionComponent<ITracesToolbarOperationsP
   );
 
   const filter = (
-    e: React.ChangeEvent<HTMLInputElement>,
+    e: React.ChangeEvent<HTMLInputElement> | null,
+    value: string,
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
   ): React.ReactElement<any, string | React.JSXElementConstructor<any>>[] => {
-    if (e && e.target && e.target.value && data) {
+    if (value && data) {
       return data
-        .filter((item) => item.includes(e.target.value))
+        .filter((item) => item.includes(value))
         .map((item, index) => <SelectOption key={index} value={item} />);
     } else {
       if (data) {