+++ plugins/opsgenie/src/components/panel/details/incident/Timeline.tsx
@@ -42,7 +42,6 @@ const Details: React.FunctionComponent<IDetailsProps> = ({ name, id }: IDetailsP
         throw err;
       }
     },
-    { keepPreviousData: true },
   );
 
   if (isLoading) {