+++ plugins/prometheus/src/components/panel/Sparkline.tsx
@@ -1,13 +1,13 @@
 import { Alert, AlertActionLink, AlertVariant } from '@patternfly/react-core';
 import { QueryObserverResult, useQuery } from 'react-query';
-import { ResponsiveLineCanvas, Serie } from '@nivo/line';
 import React from 'react';
+import { ResponsiveLineCanvas } from '@nivo/line';
 
+import { IPanelOptions, ISeries } from '../../utils/interfaces';
 import { IPluginTimes, PluginCard } from '@kobsio/plugin-core';
-import { convertMetrics, getMappingValue } from '../../utils/helpers';
+import { convertMetrics, getMappingValue, roundNumber } from '../../utils/helpers';
 import Actions from './Actions';
 import { COLOR_SCALE } from '../../utils/colors';
-import { IPanelOptions } from '../../utils/interfaces';
 
 interface ISpakrlineProps {
   name: string;
@@ -28,7 +28,7 @@ export const Spakrline: React.FunctionComponent<ISpakrlineProps> = ({
   times,
   options,
 }: ISpakrlineProps) => {
-  const { isError, isFetching, error, data, refetch } = useQuery<Serie[], Error>(
+  const { isError, isFetching, error, data, refetch } = useQuery<ISeries, Error>(
     ['prometheus/metrics', name, options.queries, times],
     async () => {
       try {
@@ -48,10 +48,10 @@ export const Spakrline: React.FunctionComponent<ISpakrlineProps> = ({
         const json = await response.json();
 
         if (response.status >= 200 && response.status < 300) {
-          if (json) {
-            return convertMetrics(json).series;
+          if (json && json.metrics) {
+            return convertMetrics(json.metrics, json.startTime, json.endTime, json.min, json.max);
           } else {
-            return [];
+            return { endTime: times.timeEnd, labels: {}, max: 0, min: 0, series: [], startTime: times.timeStart };
           }
         } else {
           if (json.error) {
@@ -70,14 +70,20 @@ export const Spakrline: React.FunctionComponent<ISpakrlineProps> = ({
   // Determine the label which should be shown above the chart. This is the last value in first metric of the returned
   // data or a value from the user specified mappings.
   let label = 'N/A';
-  if (data && data.length > 0) {
+  if (options.queries && Array.isArray(options.queries) && options.queries.length > 0 && options.queries[0].label) {
+    if (data && data.labels && data.labels.hasOwnProperty('0-0')) {
+      label = data.labels['0-0'];
+    }
+  } else if (data && data.series && data.series.length > 0) {
     if (options.mappings && Object.keys(options.mappings).length > 0) {
-      label = getMappingValue(data[0].data[data[0].data.length - 1].y, options.mappings);
+      label = getMappingValue(data.series[0].data[data.series[0].data.length - 1].y, options.mappings);
     } else {
       label =
-        data[0].data[data[0].data.length - 1].y === null
+        data.series[0].data[data.series[0].data.length - 1].y === null
           ? 'N/A'
-          : `${data[0].data[data[0].data.length - 1].y} ${options.unit ? options.unit : ''}`;
+          : `${roundNumber(data.series[0].data[data.series[0].data.length - 1].y as number)} ${
+              options.unit ? options.unit : ''
+            }`;
     }
   }
 
@@ -94,7 +100,7 @@ export const Spakrline: React.FunctionComponent<ISpakrlineProps> = ({
           title="Could not get metrics"
           actionLinks={
             <React.Fragment>
-              <AlertActionLink onClick={(): Promise<QueryObserverResult<Serie[], Error>> => refetch()}>
+              <AlertActionLink onClick={(): Promise<QueryObserverResult<ISeries, Error>> => refetch()}>
                 Retry
               </AlertActionLink>
             </React.Fragment>
@@ -102,7 +108,7 @@ export const Spakrline: React.FunctionComponent<ISpakrlineProps> = ({
         >
           <p>{error?.message}</p>
         </Alert>
-      ) : data ? (
+      ) : data && data.series ? (
         <React.Fragment>
           <div style={{ height: '100%', position: 'relative' }}>
             <div style={{ fontSize: '24px', position: 'absolute', textAlign: 'center', top: '31px', width: '100%' }}>
@@ -111,7 +117,7 @@ export const Spakrline: React.FunctionComponent<ISpakrlineProps> = ({
             <ResponsiveLineCanvas
               colors={COLOR_SCALE[0]}
               curve="monotoneX"
-              data={data}
+              data={data.series}
               enableArea={true}
               enableGridX={false}
               enableGridY={false}
@@ -119,7 +125,7 @@ export const Spakrline: React.FunctionComponent<ISpakrlineProps> = ({
               isInteractive={false}
               lineWidth={1}
               margin={{ bottom: 0, left: 0, right: 0, top: 0 }}
-              xScale={{ type: 'time' }}
+              xScale={{ max: new Date(data.endTime), min: new Date(data.startTime), type: 'time' }}
               yScale={{ stacked: false, type: 'linear' }}
             />
           </div>